updated notebook to perform numpy-based grid rf computations

+++ prfpy/timecourse.py
@@ -16,7 +16,7 @@ def convolve_stimulus_dm(stimulus, hrf):
         contains kernel for convolution
 
     """
-    hrf_shape = np.ones(len(stimulus.shape))
+    hrf_shape = np.ones(len(stimulus.shape), dtype=np.int)
     hrf_shape[-1] = hrf.shape[0]
 
     return signal.fftconvolve(stimulus, hrf.reshape(hrf_shape), mode='full', axes=(-1))[..., :stimulus.shape[-1]]
